// SPA/caams-portal/src/app/auth/asset-model/model/assetModel.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AssetModel { 
    assetCategoryOid: string;
    assetModelName: string;
    assetModelOid: string;
    assetModelSpecificationJson?: string;
    assetModelUrl?: string;
    createdBy: string;
    createdOn: Date;
    id?: number;
    manufacturerOid: string;
    organizationOid: string;
    subcategoryOid: string;
    updatedBy?: string;
    updatedOn?: Date;
}